requirements.txt
absl-py==2.3.1
anyio==4.11.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
Cython==3.2.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.7
fsspec==2025.10.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multitasking==0.0.12
namex==0.1.0
networkx==3.5
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251201
patsy==1.0.2
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
pmdarima==2.1.1
protobuf==6.33.1
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
statsmodels==0.14.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
torch==2.9.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
yfinance==0.2.66